fix for https://github.com/pycaret/pycaret/issues/1591

+++ pycaret/internal/pycaret_experiment/time_series_experiment.py
@@ -2771,7 +2771,8 @@ class TimeSeriesExperiment(_SupervisedExperiment):
                 if result.name is None:
                     result.name = self.y.name
 
-        result = result.round(round)
+        # Converting to float since rounding does not support int
+        result = result.astype(float).round(round)
 
         if isinstance(result.index, pd.DatetimeIndex):
             result.index = (
